depvet/registry/state.py
"""Registry polling state persistence."""
from __future__ import annotations
import yaml
from pathlib import Path
from typing import Any, Optional


class RegistryState:
    def __init__(self, path: str = "./depvet_state.yaml"):
        self.path = Path(path)
        self._data: dict[str, Any] = {}
        self._load()

    def _load(self) -> None:
        if self.path.exists():
            with open(self.path) as f:
                self._data = yaml.safe_load(f) or {}

    def save(self) -> None:
        self.path.parent.mkdir(parents=True, exist_ok=True)
        with open(self.path, "w") as f:
            yaml.dump(self._data, f, default_flow_style=False)

    def get(self, key: str, default: Any = None) -> Any:
        return self._data.get(key, default)

    def set(self, key: str, value: Any) -> None:
        self._data[key] = value
        self.save()
